refactor: split main entrypoint for greater clarity
