+++ cmd/server/main.go
@@ -2,10 +2,7 @@ package main
 
 import (
 	"context"
-	"io/fs"
-	"net"
 	"net/http"
-	"net/rpc"
 	"os"
 	"os/signal"
 	"syscall"
@@ -14,7 +11,6 @@ import (
 	"github.com/energietransitie/twomes-backoffice-api/handlers"
 	"github.com/energietransitie/twomes-backoffice-api/repositories"
 	"github.com/energietransitie/twomes-backoffice-api/services"
-	"github.com/energietransitie/twomes-backoffice-api/swaggerdocs"
 	"github.com/energietransitie/twomes-backoffice-api/twomes/authorization"
 	"golang.org/x/sync/errgroup"
 
@@ -24,57 +20,9 @@ import (
 )
 
 const (
-	Day = time.Hour * 24
+	shutdownTimeout    = 30 * time.Second
+	preRenewalDuration = 12 * time.Hour
 )
-
-const (
-	shutdownTimeout     = 30 * time.Second
-	preRenewalDuration  = 12 * time.Hour
-	defaultDownloadTime = "04h00s"
-)
-
-// Configuration holds all the configuration for the server.
-type Configuration struct {
-	DatabaseDSN       string
-	BaseURL           string
-	downloadStartTime time.Time
-}
-
-func getConfiguration() Configuration {
-	dsn, ok := os.LookupEnv("TWOMES_DSN")
-	if !ok {
-		logrus.Fatal("TWOMES_DSN was not set")
-	}
-
-	baseURL, ok := os.LookupEnv("TWOMES_BASE_URL")
-	if !ok {
-		logrus.Fatal("TWOMES_BASE_URL was not set")
-	}
-
-	downloadTime, ok := os.LookupEnv("TWOMES_DOWNLOAD_TIME")
-	if !ok {
-		logrus.Warning("TWOMES_DOWNLOAD_TIME was not set. defaulting to", defaultDownloadTime)
-		downloadTime = defaultDownloadTime
-	}
-
-	duration, err := time.ParseDuration(downloadTime)
-	if err != nil {
-		logrus.Fatal(err)
-	}
-
-	downloadStartTime := time.Now().Truncate(Day)
-	downloadStartTime = downloadStartTime.Add(duration)
-	// If time is in the past, add 1 day.
-	if downloadStartTime.Before(time.Now()) {
-		downloadStartTime = downloadStartTime.Add(Day)
-	}
-
-	return Configuration{
-		DatabaseDSN:       dsn,
-		BaseURL:           baseURL,
-		downloadStartTime: downloadStartTime,
-	}
-}
 
 func main() {
 	config := getConfiguration()
@@ -227,37 +175,4 @@ func listenAndServe(ctx context.Context, server *http.Server) error {
 		return err
 	}
 	return nil
-}
-
-func setupSwaggerDocs(r *chi.Mux, baseURL string) {
-	swaggerUI, err := fs.Sub(swaggerdocs.StaticFiles, "swagger-ui")
-	if err != nil {
-		logrus.Fatal(err)
-	}
-
-	docsHandler, err := handlers.NewDocsHandler(swaggerdocs.StaticFiles, baseURL)
-	if err != nil {
-		logrus.Fatal(err)
-	}
-
-	r.Method("GET", "/openapi.yml", handlers.Handler(docsHandler.OpenAPISpec))                        // Serve openapi.yml
-	r.Method("GET", "/docs/*", http.StripPrefix("/docs/", http.FileServer(http.FS(swaggerUI))))       // Serve static files.
-	r.Method("GET", "/docs", handlers.Handler(docsHandler.RedirectDocs(http.StatusMovedPermanently))) // Redirect /docs to /docs/
-	r.Method("GET", "/", handlers.Handler(docsHandler.RedirectDocs(http.StatusSeeOther)))             // Redirect / to /docs/
-}
-
-func setupAdminRPCHandler(adminHandler *handlers.AdminHandler) {
-
-	rpc.Register(adminHandler)
-	rpc.HandleHTTP()
-
-	listener, err := net.Listen("tcp4", "127.0.0.1:8081")
-	if err != nil {
-		logrus.Fatal(err)
-	}
-
-	err = http.Serve(listener, nil)
-	if err != nil {
-		logrus.Fatal(err)
-	}
 }
